(1) Compile fixes for GCC 9.1 (mingw-64) on Windows. (2) Upping MSVC warning level to 4 (3) Various level 4 warning fixes for MSVC

+++ source/Audio/Sound.h
@@ -9,5 +9,5 @@ uint32_t S_StartSound(
     const bool bStopOtherInstances = false
 ) noexcept;
 
-void S_StartSong(const uint8_t musicId) noexcept;
+void S_StartSong(const uint32_t musicId) noexcept;
 void S_StopSong() noexcept;
